backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pyserial>=3.5
aiofiles>=24.1.0
python-multipart>=0.0.12
